requirements.txt
streamlit>=1.32.0
requests>=2.31.0
pdfplumber>=0.10.3
python-docx>=1.1.0
plotly>=5.20.0
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0